# server/src/main/resources/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [{
    "url": "http://10.16.6.17:8080",
    "description": "Generated server url"
  }],
  "paths": {
    "/api/users/{userId}/favorite/{playlistId}": {
      "get": {
        "tags": ["user-controller"],
        "operationId": "isFavoritePlaylistOfUser",
        "parameters": [{
          "name": "playlistId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }, {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["user-controller"],
        "operationId": "addToFavorites",
        "parameters": [{
          "name": "playlistId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }, {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": ["user-controller"],
        "operationId": "removeFromFavorites",
        "parameters": [{
          "name": "playlistId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }, {
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "tags": ["user-controller"],
        "operationId": "getById",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutput"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["user-controller"],
        "operationId": "updateUser",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/users/{userId}/playlists": {
      "get": {
        "tags": ["user-controller"],
        "operationId": "getUserFavouritePlaylists",
        "parameters": [{
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }, {
          "name": "offset",
          "in": "query",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }, {
          "name": "limit",
          "in": "query",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistOutput"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/users/search/{search}": {
      "get": {
        "tags": ["user-controller"],
        "operationId": "searchForUsers",
        "parameters": [{
          "name": "search",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserOutput"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/user/{userId}": {
      "get": {
        "tags": ["song-controller"],
        "operationId": "getAllSongsByUser",
        "parameters": [{
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SongEntity"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/playlists/users/{user-id}/{playlist-name}": {
      "get": {
        "tags": ["playlist-controller"],
        "operationId": "getPlaylistByNameAndUserId",
        "parameters": [{
          "name": "user-id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }, {
          "name": "playlist-name",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/playlists/user/{userId}": {
      "get": {
        "tags": ["playlist-controller"],
        "operationId": "getAllPlaylistsByUser",
        "parameters": [{
          "name": "userId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }, {
          "name": "offset",
          "in": "query",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }, {
          "name": "limit",
          "in": "query",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }
        }],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlaylistOutput"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlaylistOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isAllSongs": {
            "type": "boolean"
          },
          "createDate": {
            "type": "string",
            "format": "date"
          },
          "type": {
            "type": "string",
            "enum": ["PRIVATE", "PUBLIC"]
          },
          "imageUrl": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "UserInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "oldPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "isPremium": {
            "type": "boolean"
          },
          "leftSongs": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UserOutput": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "isPremium": {
            "type": "boolean"
          },
          "leftSongs": {
            "type": "integer",
            "format": "int32"
          },
          "role": {
            "type": "string",
            "enum": ["USER", "ADMIN"]
          }
        }
      },
...
      "SongEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "releaseYear": {
            "type": "integer",
            "format": "int32"
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "uploadDate": {
            "type": "string",
            "format": "date"
          },
          "imageUrl": {
            "type": "string"
          },
          "cloudUrl": {
            "type": "string"
          }
        }
      }
    }
  }
}